# docs/swagger.yaml
definitions:
  app.Pager:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total_rows:
        type: integer
    type: object
  errcode.Error:
    type: object
  model.Tag:
    properties:
      created_by:
        type: string
      created_on:
        type: integer
      deleted_on:
        type: integer
      id:
        type: integer
      is_del:
        type: integer
      modified_by:
        type: string
      modified_on:
        type: integer
      name:
        type: string
      state:
        type: string
    type: object
  model.TagSwagger:
    properties:
      list:
        items:
          $ref: '#/definitions/model.Tag'
        type: array
      pager:
        $ref: '#/definitions/app.Pager'
        type: object
    type: object
info:
  contact: {}
  description: blog service demo
  license: {}
  title: blog service demo
  version: "1.0"
paths:
  /api/v1/tags:
    get:
      parameters:
      - description: tag name
        in: query
        maxLength: 100
        name: name
        type: string
      - default: 1
        description: state
        enum:
        - 0
        - 1
        in: query
        name: state
        type: integer
      - description: page number
        in: query
        name: page
        type: integer
      - description: number of every page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: request success
          schema:
            $ref: '#/definitions/model.TagSwagger'
        "400":
          description: request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: get tags
    post:
      parameters:
      - description: tag name
        in: query
        maxLength: 100
        name: name
        type: string
      - default: 1
        description: state
        enum:
        - 0
        - 1
        in: query
        name: state
        type: integer
      - description: creator
        in: body
        maxLength: 100
        minLength: 3
        name: created_by
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: request success
          schema:
            $ref: '#/definitions/model.Tag'
        "400":
          description: request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: create tag
  /api/v1/tags/{id}:
    delete:
      parameters:
      - description: tag ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: request success
          schema:
            $ref: '#/definitions/model.Tag'
        "400":
          description: request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: delete tag
    put:
      parameters:
      - description: tag ID
        in: path
        name: id
        required: true
        type: integer
      - description: tag name
        in: query
        maxLength: 100
        name: name
        type: string
      - default: 1
        description: state
        enum:
        - 0
        - 1
        in: query
        name: state
        type: integer
      - description: modifier
        in: body
        maxLength: 100
        minLength: 3
        name: modified_by
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: request success
          schema:
            $ref: '#/definitions/model.Tag'
        "400":
          description: request error
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: internal error
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: update tag
swagger: "2.0"